Update pointCountChanged signal
Updated the pointCountChanged signal to pass a size_t parameter, which
matches the type of the underlying container.

+++ gui/GroundControlPointsHelper.h
@@ -75,7 +75,7 @@ public slots:
   void enableWidgets(bool);
 
 signals:
-  void pointCountChanged(int);
+  void pointCountChanged(size_t);
   void pointAdded(kwiver::vital::ground_control_point_id_t);
   void pointRemoved(kwiver::vital::ground_control_point_id_t);
   void pointChanged(kwiver::vital::ground_control_point_id_t);
